pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools", "setuptools-scm" ]

[project]
name = "djangocms-automation"
description = "Automation tools for django CMS projects"
readme = "README.md"
license = "BSD-3-Clause"
authors = [
  { name = "Fabian Braun", email = "fsbraun@gmx.de" },
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
  "Framework :: Django CMS :: 4.1",
  "Framework :: Django CMS :: 5.0",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]
dependencies = [
  "django>=4.2",
  "django-cms>=4.1",
  "jsonschema>=4.22",
]

optional-dependencies.test = [
  "coverage",
  "pytest",
  "pytest-django",
]

urls.Homepage = "https://github.com/fsbraun/djangocms-automation"
urls.Issues = "https://github.com/fsbraun/djangocms-automation/issues"

[tool.setuptools.dynamic]
version = { attr = "djangocms_automation.__version__" }

[tool.setuptools.packages.find]
where = [ "." ]
include = [ "djangocms_automation*" ]

[tool.setuptools.package-data]
"*" = [
  "templates/**/*.html",
  "static/**/*",
  "locale/**/*",
  "*.po",
  "*.mo",
]

[tool.ruff]
target-version = "py39"
line-length = 119
exclude = [ "djangocms_automation/migrations" ]
lint.ignore = [  ]
lint.mccabe.max-complexity = 5

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["--tb=short", "--strict-markers", "--nomigrations"]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "django_db: mark test as requiring database access",
]

[tool.coverage.run]
source = [
  "djangocms_automation",
]
omit = [
  "djangocms_automation/migrations/*",
]
